добавил реализацию кнопок + тесты

+++ src/main/java/com/malex/service/Operation.java
@@ -2,9 +2,11 @@ package com.malex.service;
 
 import com.malex.model.enums.OperationName;
 
+import java.math.BigDecimal;
+
 public interface Operation {
 
     OperationName getOperationName();
 
-    int execute(int numberOne, int numberTwo);
+    BigDecimal execute(BigDecimal numberOne, BigDecimal numberTwo);
 }
